Renamed three_player_rationale to introspective_rationale, addressed some nit comments in open PR 46
