+++ python/interpret_text/common/utils_introspective_rationale.py
