+++ test/unit/test_introspective_rationale_explainer.py
@@ -7,11 +7,11 @@ import os
 import torch
 import pandas as pd
 
-from interpret_text.three_player_introspective.three_player_introspective_explainer import (
-    ThreePlayerIntrospectiveExplainer
+from interpret_text.introspective_rationale.introspective_rationale_explainer import (
+    IntrospectiveRationaleExplainer
 )
 from interpret_text.common.dataset.utils_sst2 import load_sst2_pandas_df
-from interpret_text.common.utils_three_player import GlovePreprocessor, ModelArguments, load_glove_embeddings
+from interpret_text.common.utils_introspective_rationale import GlovePreprocessor, ModelArguments, load_glove_embeddings
 
 
 class TestThreePlayerExplainer(object):
